rename settingsValidator to settingsFactory
