+++ src/__tests__/browserSuites/fetch-specific-splits.spec.js
@@ -1,4 +1,4 @@
-import { SplitFactory } from '../../index';
+import { SplitFactory } from '../../';
 import { splitFilters, queryStrings, groupedFilters } from '../mocks/fetchSpecificSplits';
 
 const baseConfig = {
