+++ src/__tests__/browserSuites/push-initialization-retries.spec.js
@@ -7,8 +7,8 @@ import mySegmentsNicolasMock from '../mocks/mysegments.nicolas@split.io.json';
 
 import { nearlyEqual, url } from '../testUtils';
 import EventSourceMock, { setMockListener } from '../testUtils/eventSourceMock';
-import { SplitFactory } from '../../index';
-import { settingsValidator } from '../../settings';
+import { SplitFactory } from '../../';
+import { settingsFactory } from '../../settings';
 
 const baseUrls = {
   sdk: 'https://sdk.push-initialization-retries/api',
@@ -34,7 +34,7 @@ const config = {
   streamingEnabled: true,
   // debug: true,
 };
-const settings = settingsValidator(config);
+const settings = settingsFactory(config);
 
 /**
  * Sequence of calls:
