+++ src/__tests__/browserSuites/push-refresh-token.spec.js
@@ -10,8 +10,8 @@ import { nearlyEqual, url } from '../testUtils';
 import EventSourceMock, { setMockListener } from '../testUtils/eventSourceMock';
 window.EventSource = EventSourceMock;
 
-import { SplitFactory } from '../../index';
-import { settingsValidator } from '../../settings';
+import { SplitFactory } from '../../';
+import { settingsFactory } from '../../settings';
 
 const userKey = 'nicolas@split.io';
 
@@ -32,7 +32,7 @@ const config = {
   },
   // debug: true,
 };
-const settings = settingsValidator(config);
+const settings = settingsFactory(config);
 
 const MILLIS_CONNDELAY = 500;
 const MILLIS_REFRESH_TOKEN = 1000;
