+++ src/__tests__/browserSuites/push-synchronization-retries.spec.js
@@ -17,8 +17,8 @@ import { Backoff } from '@splitsoftware/splitio-commons/src/utils/Backoff';
 import EventSourceMock, { setMockListener } from '../testUtils/eventSourceMock';
 window.EventSource = EventSourceMock;
 
-import { SplitFactory } from '../../index';
-import { settingsValidator } from '../../settings';
+import { SplitFactory } from '../../';
+import { settingsFactory } from '../../settings';
 
 const userKey = 'nicolas@split.io';
 const otherUserKeySync = 'marcio@split.io';
@@ -37,7 +37,7 @@ const config = {
   streamingEnabled: true,
   // debug: true,
 };
-const settings = settingsValidator(config);
+const settings = settingsFactory(config);
 
 const MILLIS_SSE_OPEN = 100;
