+++ src/__tests__/browserSuites/user-consent.spec.js
@@ -1,5 +1,5 @@
 import sinon from 'sinon';
-import { SplitFactory } from '../../index';
+import { SplitFactory } from '../../';
 import { nearlyEqual, url } from '../testUtils';
 import { triggerPagehideEvent } from '../testUtils/browser';
