+++ src/__tests__/consumer/browser_consumer.spec.js
@@ -5,7 +5,7 @@ import { SDK_NOT_READY, EXCEPTION } from '@splitsoftware/splitio-commons/src/uti
 import { applyOperations } from './wrapper-commands';
 import { nearlyEqual } from '../testUtils';
 
-import { SplitFactory, PluggableStorage } from '../../index';
+import { SplitFactory, PluggableStorage } from '../../';
 
 const expectedSplitName = 'hierarchical_splits_testing_on';
 const expectedSplitView = { name: 'hierarchical_splits_testing_on', trafficType: 'user', killed: false, changeNumber: 1487277320548, treatments: ['on', 'off'], configs: {} };
