+++ src/__tests__/errorCatching/browser.spec.js
@@ -7,10 +7,10 @@ import splitChangesMock1 from '../mocks/splitChanges.since.-1.till.1500492097547
 import mySegmentsMock from '../mocks/mySegmentsEmpty.json';
 import splitChangesMock2 from '../mocks/splitChanges.since.1500492097547.till.1500492297547.json';
 import splitChangesMock3 from '../mocks/splitChanges.since.1500492297547.json';
-import { SplitFactory, InLocalStorage } from '../../index';
-import { settingsValidator } from '../../settings';
+import { SplitFactory, InLocalStorage } from '../../';
+import { settingsFactory } from '../../settings';
 
-const settings = settingsValidator({
+const settings = settingsFactory({
   core: {
     authorizationKey: '<fake-token>'
   },
