+++ src/__tests__/gaIntegration/both-integrations.spec.js
@@ -1,5 +1,5 @@
-import { SplitFactory, GoogleAnalyticsToSplit, SplitToGoogleAnalytics } from '../../index';
-import { settingsValidator } from '../../settings';
+import { SplitFactory, GoogleAnalyticsToSplit, SplitToGoogleAnalytics } from '../../';
+import { settingsFactory } from '../../settings';
 import { gaSpy, gaTag } from './gaTestUtils';
 import includes from 'lodash/includes';
 import { DEBUG } from '@splitsoftware/splitio-commons/src/utils/constants';
@@ -21,7 +21,7 @@ const config = {
     impressionsMode: DEBUG,
   }
 };
-const settings = settingsValidator(config);
+const settings = settingsFactory(config);
 
 export default function (fetchMock, assert) {
