+++ src/__tests__/offline/browser.spec.js
@@ -3,10 +3,10 @@ import sinon from 'sinon';
 import fetchMock from '../testUtils/fetchMock';
 import { url } from '../testUtils';
 import { SplitFactory, InLocalStorage } from '../../full';
-import { SplitFactory as SplitFactorySlim, LocalhostFromObject } from '../../index';
-import { settingsValidator } from '../../settings';
+import { SplitFactory as SplitFactorySlim, LocalhostFromObject } from '../../';
+import { settingsFactory } from '../../settings';
 
-const settings = settingsValidator({ core: { key: 'facundo@split.io' } });
+const settings = settingsFactory({ core: { key: 'facundo@split.io' } });
 
 const spySplitChanges = sinon.spy();
 const spySegmentChanges = sinon.spy();
