+++ src/full/splitFactory.ts
@@ -1,4 +1,4 @@
-import { settingsValidator } from '../settings/full';
+import { settingsFactory } from '../settings/full';
 import { getModules } from '../platform/getModules';
 import { sdkFactory } from '@splitsoftware/splitio-commons/src/sdkFactory/index';
 import { ISdkFactoryParams } from '@splitsoftware/splitio-commons/src/sdkFactory/types';
@@ -6,6 +6,7 @@ import { getFetch } from '../platform/getFetchFull';
 import { getEventSource } from '../platform/getEventSource';
 import { EventEmitter } from '@splitsoftware/splitio-commons/src/utils/MinEvents';
 import { now } from '@splitsoftware/splitio-commons/src/utils/timeTracker/now/browser';
+import { IBrowserSettings } from '../../types/splitio';
 
 const platform = { getFetch, getEventSource, EventEmitter, now };
 
@@ -18,8 +19,8 @@ const platform = { getFetch, getEventSource, EventEmitter, now };
  * caution since, unlike `config`, this param is not validated neither considered part of the public API.
  * @throws Will throw an error if the provided config is invalid.
  */
-export function SplitFactory(config: any, __updateModules?: (modules: ISdkFactoryParams) => void) {
-  const settings = settingsValidator(config);
+export function SplitFactory(config: IBrowserSettings, __updateModules?: (modules: ISdkFactoryParams) => void) {
+  const settings = settingsFactory(config);
   const modules = getModules(settings, platform);
   if (__updateModules) __updateModules(modules);
   return sdkFactory(modules);
