+++ src/settings/__tests__/index.spec.ts
@@ -1,18 +1,18 @@
-import { settingsValidator as slimSettingsValidator } from '../index';
-import { settingsValidator as fullSettingsValidator } from '../index';
+import { settingsFactory as slimSettingsFactory } from '../index';
+import { settingsFactory as fullSettingsFactory } from '../full';
 
 test('SETTINGS / Consent is overwritable and "GRANTED" by default', () => {
-  [slimSettingsValidator, fullSettingsValidator].forEach((settingsValidator) => {
-    let settings = settingsValidator({});
+  [slimSettingsFactory, fullSettingsFactory].forEach((settingsFactory) => {
+    let settings = settingsFactory({});
     expect(settings.userConsent).toEqual('GRANTED'); // userConsent defaults to granted if not provided
 
-    settings = settingsValidator({ userConsent: 'INVALID-VALUE' });
+    settings = settingsFactory({ userConsent: 'INVALID-VALUE' });
     expect(settings.userConsent).toEqual('GRANTED'); // userConsent defaults to granted if a wrong value is provided
 
-    settings = settingsValidator({ userConsent: 'UNKNOWN' });
+    settings = settingsFactory({ userConsent: 'UNKNOWN' });
     expect(settings.userConsent).toEqual('UNKNOWN'); // userConsent can be overwritten
 
-    settings = settingsValidator({ userConsent: 'declined' });
+    settings = settingsFactory({ userConsent: 'declined' });
     expect(settings.userConsent).toEqual('DECLINED'); // userConsent can be overwritten
   });
 });
